Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,7 +3,7 @@ int main()
 {
     int a,b,c;
     scanf("%d %d %d", a , b, c);
-    double s=(a+b+c)/3;
-    printf("Average: %.2lf", s);
+    float s=(a+b+c)/3;
+    printf("Average: %.2f", s);
     return 0;
 }
